+++ hazelcastmq-examples/src/main/java/org/mpilone/hazelcastmq/example/camel/CamelToCamelOneWay.java
@@ -7,6 +7,7 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.impl.DefaultCamelContext;
 import org.mpilone.hazelcastmq.camel.*;
 import org.mpilone.hazelcastmq.core.*;
+import org.mpilone.hazelcastmq.example.ExampleApp;
 import org.slf4j.*;
 
 import com.hazelcast.config.Config;
@@ -18,7 +19,7 @@ import com.hazelcast.core.*;
  *
  * @author mpilone
  */
-public class CamelToCamelOneWay {
+public class CamelToCamelOneWay extends ExampleApp {
 
   /**
    * The log for this class.
@@ -27,18 +28,15 @@ public class CamelToCamelOneWay {
       CamelToCamelOneWay.class);
 
   public static void main(String[] args) throws Exception {
-    System.setProperty("org.slf4j.simpleLogger.defaultLogLevel", "debug");
-    System.setProperty("org.slf4j.simpleLogger.showDateTime", "true");
-    System.setProperty("org.slf4j.simpleLogger.log.com.hazelcast", "info");
-
-    new CamelToCamelOneWay();
+    CamelToCamelOneWay app = new CamelToCamelOneWay();
+    app.runExample();
   }
 
-  public CamelToCamelOneWay() throws Exception {
+  @Override
+  public void start() throws Exception {
 
     // Create a Hazelcast instance.
     Config config = new Config();
-    config.setProperty("hazelcast.logging.type", "slf4j");
     config.getNetworkConfig().getJoin().getMulticastConfig().setEnabled(false);
     HazelcastInstance hazelcast = Hazelcast.newHazelcastInstance(config);