+++ hazelcastmq-examples/src/main/java/org/mpilone/hazelcastmq/example/spring/tx/support/hz/DemoQueueReader.java
@@ -1,5 +1,5 @@
 
-package org.mpilone.hazelcastmq.example.spring.tx.support;
+package org.mpilone.hazelcastmq.example.spring.tx.support.hz;
 
 import java.util.concurrent.TimeUnit;