+++ hazelcastmq-jms/src/main/java/org/mpilone/hazelcastmq/jms/IdGenerator.java
@@ -1,5 +1,5 @@
 package org.mpilone.hazelcastmq.jms;
 
-interface IdGenerator {
+public interface IdGenerator {
   public String newId();
 }